requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-embed-video==1.3.3
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-mptt==0.11.0
django-pagedown==2.1.2
django-taggit==1.2.0
humanize==0.5.1
idna==2.9
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
django-admin-honeypot==1.1.0